Prevent compilation error when comparing ints to unsigned ints

+++ src/thinker/field.c
@@ -65,8 +65,8 @@ bool isValidMove(Field* field, char own, unsigned int x, unsigned int y) {
     char other = own == 'W' ? 'B' : 'W';
 
     // vertical check ^
-    int found = -1;
-    for(int i = (int)y; i >= 0; i--) {
+    long int found = -1;
+    for(unsigned int i = y; i >= 0; i--) {
         char column = getColumn(field, x, i);
 
         if(column == '*' && i != y) {
@@ -83,7 +83,7 @@ bool isValidMove(Field* field, char own, unsigned int x, unsigned int y) {
     }
     found = -1;
 
-    for(int i = (int)y; i < field->height; i++) {
+    for(unsigned int i = y; i < field->height; i++) {
         char column = getColumn(field, x, i);
 
         if(column == '*' && i != y)
@@ -99,7 +99,7 @@ bool isValidMove(Field* field, char own, unsigned int x, unsigned int y) {
     }
     found = -1;
 
-    for(int j = (int)x; j >= 0; j--) {
+    for(unsigned int j = x; j >= 0; j--) {
         char column = getColumn(field, j, y);
 
         if(column == '*' && j != x)
@@ -116,7 +116,7 @@ bool isValidMove(Field* field, char own, unsigned int x, unsigned int y) {
     found = -1;
 
     // horizontal check >
-    for(int j = (int)x; j < (int)field->width; j++) {
+    for(unsigned int j = x; j < field->width; j++) {
         char column = getColumn(field, j, y);
 
         if(column == '*' && j != x)
@@ -133,7 +133,7 @@ bool isValidMove(Field* field, char own, unsigned int x, unsigned int y) {
     found = -1;
 
     // diagonal check <^
-    for(int i = (int)y, j = (int)x; i >= 0 && j >= 0; j--, i--) {
+    for(unsigned int i = y, j = x; i >= 0 && j >= 0; j--, i--) {
         char column = getColumn(field, j, i);
 
         if(column == '*' && i != y)
@@ -149,7 +149,7 @@ bool isValidMove(Field* field, char own, unsigned int x, unsigned int y) {
     }
 
     // diagonal check |>
-    for(int i = (int)y, j = (int)x; i < field->height && j < field->width; j++, i++) {
+    for(unsigned int i = y, j = x; i < field->height && j < field->width; j++, i++) {
         char column = getColumn(field, j, i);
 
         if(column == '*' && i != y)
